requirements.txt
langchain
huggingface_hub
langchain_community
transformers
python-dotenv
tokenizers
streamlit
